+++ src/NexGen.MediatR.Extensions.Caching.Redis/Constants/CacheKeys.cs
@@ -5,4 +5,5 @@ internal static class CacheKeys
     internal const string CacheTypesKey = "NexGen.MediatR.Caching:Container:CacheTypes";
     internal const string CacheTagsKey = "NexGen.MediatR.Caching:Container:CacheTags";
     internal const string RequestResponseTypesKey = "NexGen.MediatR.Caching:Container:RequestResponseTypes";
+    internal const string EvictionChannel = "NexGen.MediatR.Caching:Evict";
 }