+++ src/NexGen.MediatR.Extensions.Caching/Configurations/RequestOutputCacheConfiguration.cs
@@ -34,6 +34,8 @@ public static class RequestOutputCacheConfiguration
     /// <returns>The updated <see cref="IServiceCollection"/>.</returns>
     private static IServiceCollection AddMediatROutputCache(this IServiceCollection services)
     {
-        return services.AddScoped(typeof(IPipelineBehavior<,>), typeof(RequestOutputCacheBehavior<,>));
+        services.AddScoped(typeof(IPipelineBehavior<,>), typeof(RequestOutputCacheBehavior<,>));
+        services.AddScoped(typeof(IPipelineBehavior<,>), typeof(RequestOutputCacheEvictBehavior<,>));
+        return services;
     }
 }